refactor: disable swipe of page view controller & add action to change pages
